Add a common & shared karma config
